Move checksum function to package checksum
ref #1364
