+++ portal/src/graphql/portal/schema.graphql
@@ -114,7 +114,9 @@ type AppListItem {
 
 """Resource file for an app"""
 type AppResource {
-  """"""
+  """
+  The checksum of the resource file. It is an opaque string that will be used to detect conflict.
+  """
   checksum: String
 
   """"""
@@ -132,7 +134,9 @@ type AppResource {
 
 """Update to resource file."""
 input AppResourceUpdate {
-  """Checksum of the resource file. Set to null to remove it."""
+  """
+  The checksum of the original resource file. If provided, it will be used to detect conflict.
+  """
   checksum: String
 
   """New data of the resource file. Set to null to remove it."""
@@ -1002,7 +1006,9 @@ input UpdateAppInput {
   """authgear.yaml in JSON."""
   appConfig: AppConfig
 
-  """Checksum of the app config. Set to null to remove it."""
+  """
+  The checksum of appConfig. If provided, it will be used to detect conflict.
+  """
   appConfigChecksum: String
 
   """App ID to update."""
@@ -1012,7 +1018,7 @@ input UpdateAppInput {
   secretConfigUpdateInstructions: SecretConfigUpdateInstructionsInput
 
   """
-  Checksum of the secret config update instructions. Set to null to remove it.
+  The checksum of secretConfig. If provided, it will be used to detect conflict.
   """
   secretConfigUpdateInstructionsChecksum: String
